@sanderegg review: rename module
