+++ packages/models-library/src/models_library/services.py
@@ -8,16 +8,16 @@ from .boot_options import BootOption, BootOptions
 from .emails import LowerCaseEmailStr
 from .services_authoring import Author, Badge
 from .services_base import ServiceBase, ServiceKeyVersion
-from .services_constants import ANY_FILETYPE, LATEST_INTEGRATION_VERSION
-from .services_constrained_types import (
+from .services_enums import ServiceType
+from .services_io import BaseServiceIOModel, ServiceInput, ServiceOutput
+from .services_regex import ANY_FILETYPE, LATEST_INTEGRATION_VERSION
+from .services_types import (
     DynamicServiceKey,
     RunID,
     ServiceKey,
     ServicePortKey,
     ServiceVersion,
 )
-from .services_enums import ServiceType
-from .services_io import BaseServiceIOModel, ServiceInput, ServiceOutput
 
 # NOTE: these asserts are here to avoid pre-commit to prune these imports
 assert BaseServiceIOModel  # nsoec
