+++ packages/models-library/src/models_library/services_metadata.py
@@ -4,9 +4,9 @@ from typing import Any, ClassVar
 from pydantic import Field, HttpUrl
 
 from .services_base import ServiceBase
-from .services_constants import LATEST_INTEGRATION_VERSION
-from .services_constrained_types import DynamicServiceKey, ServiceKey, ServiceVersion
 from .services_enums import ServiceType
+from .services_regex import LATEST_INTEGRATION_VERSION
+from .services_types import DynamicServiceKey, ServiceKey, ServiceVersion
 
 assert DynamicServiceKey  # nosec
 assert LATEST_INTEGRATION_VERSION  # nosec
