+++ packages/service-integration/src/service_integration/osparc_config.py
@@ -25,7 +25,7 @@ from models_library.service_settings_labels import (
 )
 from models_library.service_settings_nat_rule import NATRule
 from models_library.services import BootOptions, ServiceMetaDataPublished, ServiceType
-from models_library.services_constants import (
+from models_library.services_regex import (
     COMPUTATIONAL_SERVICE_KEY_FORMAT,
     DYNAMIC_SERVICE_KEY_FORMAT,
 )
