+++ services/api-server/src/simcore_service_api_server/models/schemas/solvers.py
@@ -4,7 +4,7 @@ from typing import Any, ClassVar, Literal
 import packaging.version
 from models_library.basic_regex import PUBLIC_VARIABLE_NAME_RE
 from models_library.services import ServiceMetaDataPublished
-from models_library.services_constants import COMPUTATIONAL_SERVICE_KEY_RE
+from models_library.services_regex import COMPUTATIONAL_SERVICE_KEY_RE
 from packaging.version import Version
 from pydantic import BaseModel, ConstrainedStr, Extra, Field, HttpUrl
