+++ services/director-v2/src/simcore_service_director_v2/models/comp_tasks.py
@@ -11,7 +11,7 @@ from models_library.projects_nodes import InputsDict, NodeID, OutputsDict
 from models_library.projects_state import RunningState
 from models_library.resource_tracker import HardwareInfo
 from models_library.services import ServiceInputsDict, ServiceOutput, ServicePortKey
-from models_library.services_constants import SERVICE_KEY_RE
+from models_library.services_regex import SERVICE_KEY_RE
 from models_library.services_resources import BootMode
 from pydantic import BaseModel, ByteSize, Extra, Field, parse_obj_as, validator
 from pydantic.types import PositiveInt
