+++ services/director-v2/src/simcore_service_director_v2/utils/computations.py
@@ -4,7 +4,7 @@ from typing import Any
 
 from models_library.projects_state import RunningState
 from models_library.services import ServiceKeyVersion
-from models_library.services_constants import SERVICE_KEY_RE
+from models_library.services_regex import SERVICE_KEY_RE
 from models_library.users import UserID
 from pydantic import parse_obj_as
 from servicelib.utils import logged_gather
